Refactor field name validation in EntryEditor preferences, using FieldFactory, and change default set of unwanted symbols.
Replace custom validation logic with `FieldFactory.isValidFieldName`, extracting it to a method in FieldFactory and changing the logic for validation.

Previously, the validation consisted of generating citation keys using each field name defined by the user. However, citation keys are not the same as field names. The original code acknowledge that in the inline comment. Moreover, while the list of unwanted symbols for citation keys was configurable, the list of unwanted symbols for field names were not.

Thus, I refactored the validation of filed names, extracting the method FieldFactory.isValidFieldName() and using it to validate the field name. The list of unwanted symbol was also refactored, by extracting it to a public constant at FieldFactory.UNWANTED_CHARS. I do not kept that list configurable because BibTeX and Biblatex as I consider the current set as reasonable.